Restrain waitForNotificationsIfNewBlocksConnected check to strict tip
waitForNotificationsIfNewBlocksConnected() would previously return immediately
if the best block processed by the wallet was a descendant of the node's
tip. This would mean that during a reorg, an RPC could return before the
BlockDisconnected callback had been triggered. This could cause a tx
that had been included in the disconnected block to be displayed as
confirmed, for example.

Instead, change waitForNotificationsIfNewBlocksConnected() to
waitForNotificationsIfTipChanged() and only return early if the best
block processed by the wallet is the node's tip.