[watchdog] Introduce CWatchdog class
CWatchdog is a cross-layer anomalies detection module aims to
spot any anomalies at the headers reception, block processing,
net level or peer management.

If any anomaly is detected an event is triggered to inform an
application layer or any consumer of a future watchdog interface.

This code is only integrated in next commit.