[net/refactor] Rework logic to make connection counts explicit
Changing the logic into a switch statement makes it more apparent what the
logic flow is for each type of connection. This should help prevent bugs around
the counts. Having a switch statement also ensures we will get a compiler
warning if another connection type is added to the enum but not to this
conditional.