[validation/refactor] CheckSequenceLocks accepts any coinsview
Allows CheckSequenceLocks to use heights and coins from any CoinsView
and CBlockIndex provided. The typical usage would still be to create a
CCoinsViewMemPool from a pool and pass it in. This also means that
CheckSequenceLocks() doesn't need to hold the mempool lock because it
doesn't interact with it. The caller is responsible for grabbing the
mempool lock if it creates the coins view from it.